+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## Unreleased
 
-- Support for propagating errors via an `assign_errors` helper
-- Preservation of assigned errors across redirects
+- Support for propagating errors via an `assign_errors` helper ([#10](https://github.com/svycal/inertia-phoenix/issues/10))
+- Preservation of assigned errors across redirects ([#10](https://github.com/svycal/inertia-phoenix/issues/10))
+- Setup external redirects properly for Inertia requests ([#11](https://github.com/svycal/inertia-phoenix/issues/11))
 
 ## 0.4.0