+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 ## Unreleased
 
-- Support for propagating errors via an `assign_errors` helper ([#10](https://github.com/svycal/inertia-phoenix/issues/10))
-- Preservation of assigned errors across redirects ([#10](https://github.com/svycal/inertia-phoenix/issues/10))
-- Setup external redirects properly for Inertia requests ([#11](https://github.com/svycal/inertia-phoenix/issues/11))
+- Assign errors via an `assign_errors` helper ([#10](https://github.com/svycal/inertia-phoenix/issues/10))
+- Preserve assigned errors across redirects ([#10](https://github.com/svycal/inertia-phoenix/issues/10))
+- Set up external redirects properly for Inertia requests ([#11](https://github.com/svycal/inertia-phoenix/issues/11))
+- Pass CSRF tokens via cookies ([#12](https://github.com/svycal/inertia-phoenix/issues/12)) 
 - Forward flash contents across forced refreshes ([#13](https://github.com/svycal/inertia-phoenix/issues/13))
 - Automatically pass Phoenix flash data via the `flash` prop